+++ test/td_df/templates/templates_test.exs
@@ -87,6 +87,57 @@ defmodule TdDf.TemplatesTest do
       assert {:error, %Ecto.Changeset{}} = Templates.create_template(@invalid_attrs)
     end
 
+    test "create_template/1 with repeated name field in content returns error changeset" do
+      content = [
+        %{"name" => "my repeated name"},
+        %{"name" => "my name"},
+        %{"name" => "my repeated name"}
+      ]
+
+      attrs = Map.put(@valid_attrs, :content, content)
+      assert {:error, %Ecto.Changeset{errors: errors}} = Templates.create_template(attrs)
+
+      error =
+        errors
+        |> Keyword.get(:content)
+        |> elem(0)
+
+      assert error == "repeated.field"
+    end
+
+    test "create_template/1 with with existing fields and different type in another template returns error changeset" do
+      insert(:template, content: [%{"name" => "field1", "type" => "type1"}])
+
+      content = [
+        %{"name" => "field1", "type" => "typex"},
+        %{"name" => "field2", "type" => "type1"}
+      ]
+
+      attrs = Map.put(@valid_attrs, :content, content)
+      assert {:error, %Ecto.Changeset{errors: errors}} = Templates.create_template(attrs)
+
+      error =
+        errors
+        |> Keyword.get(:content)
+        |> elem(0)
+
+      assert error == "invalid.type"
+    end
+
+    test "create_template/1 with with existing fields and same type in another template creates the template" do
+      insert(:template, content: [%{"name" => "field1", "type" => "type1"}])
+
+      content = [
+        %{"name" => "field1", "type" => "type1"},
+        %{"name" => "field2", "type" => "type1"}
+      ]
+
+      attrs = Map.put(@valid_attrs, :content, content)
+      assert {:ok, %Template{} = template} = Templates.create_template(attrs)
+
+      assert template.content == content
+    end
+
     test "update_template/2 with valid data updates the template" do
       template = template_fixture()
       assert {:ok, template} = Templates.update_template(template, @update_attrs)